+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Compatibility:
 * Allow anonymous memberless Struct (@simonlevasseur).
 * Set `$!` when a `Kernel#at_exit` hook raises an exception (#3535, @andrykonchin).
 * Support `:buffer` keyword argument to `Array#pack` (#3559, @andrykonchyn).
+* Set `RbConfig::CONFIG['host_cpu']` to `arm64` on darwin platform (#3571, @andrykonchin).
 
 Performance:
 * Fix inline caching for Regexp creation from Strings (#3492, @andrykonchin, @eregon).